one dataset class per module
